requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# LLM backends
openai>=1.12.0
anthropic>=0.21.0
llama-cpp-python>=0.2.0

# RAG and vector storage
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22
sentence-transformers>=2.3.1 
faiss-cpu>=1.7.4

# NLP utilities
tiktoken>=0.5.2
huggingface-hub>=0.20.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Logging and monitoring
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
